BCStateTran: add SourceBatch, perform background sending at source
Since sending a batch can take very long time, up to seconds, and since
ST main thread must be responsive - we must do it in an async IO manner -
in the background.
Here we add a new state SendingBatch. We introduce a new struct
SourceBatch and a new function continueSendBatch(). This function
can process sending of chunks to a destination up to
maxProcessingTimeMilli. It waits for blocks from the worker thread up to
maxWaitingTimeMilli.

In addition, at the end of sending batch, the session is closed if the source
sees that this is the last batch in a cycle. This improves the time wasted
on short ST cycles (read-only replicas) and increases source availability.

Also, here we add more metrics/histograms for better visibility and
analysis of code.