BCStateTran: use class Handoff for a complete interface separation
for CRE, some nonvirtual APIs and IStateTransfer:
Interfaces can be implemented in 3 ways:
1) Atomic interface: usually a simple getter - get an atomic member
value. This one bypasses the Handoff class.
2) Via Handoff class, by pushing into incomingEventsQ_:
2a) Non-Blocking - few functions use this capability, whenever an
asynchronous code is written on the caller side. Call and parameters are
pushed and the calling thread continues immediately.
2b) Blocking - most interfaces are implemented that way. In this mode,
the caller pushes a request (a call) into the Handoff class and waits for a
reply. Blocked calls get priority over Non-blocked calls.

Comment: IReservedPages is not separated, since it is already
thread-safe.

Other changes:
1) Refactor handlers into a new file BCStateTranInterface.cpp.
2) Reduce visibility - all interfaces are public. All the rest of the
members are protected.
3) Modify bcstatetransfer_tests UT to match changes.
4) RVB Manager - completely thread-safe. reportLastAgreedPrunableBlockId
is not blocking. Remove detection mechanism, remove locks.
5) Misc minor cosmetics + refactoring.
6) New state is added to ST - to support blocking mode, ST must serve
calls during the last phase of callbacks. This phase is called
FinalizingCycle. A one-shot timer is used to wait for a worker thread to
finish calling all callbacks.